Update to Django 1.11.
There's a functional change here; instead of returning a string
in debug mode with a suitable error, we now raise an exception. This
is closer in spirit to where Django is these days (as I understand
things).

The main underlying change is to move away from using the Context
stack to building a temporary dictionary (no matter whether passed
a Context or a dict) that we can mutate as we wish. This is in
theory compatible with template engines other than DTL, eg Jinja2,
although as yet completely untested.